w3c-reports/12092014/mac-wk2/runned-tests/IndexedDB/idbcursor-direction-index-expected.txt

CONSOLE MESSAGE: line 50: DataError: DOM IDBDatabase Exception 0: The data provided does not meet requirements.

Harness Error. harness_status.status = 1 , harness_status.message = DataError: DOM IDBDatabase Exception 0: The data provided does not meet requirements.
